backend/requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8

# new
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
djangorestframework==3.12.4
idna==3.2
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
djangorestframework-simplejwt==4.7.2
celery>=5.0
redis>=3.5
drf-yasg==1.20.0
Pillow==8.3.1
django-cors-headers==3.7.0

